refactor(cli): extract parsers to helpers and consolidate tests
