+++ src-tauri/src/cli/mod.rs
@@ -1,4 +1,5 @@
 pub mod cli;
+pub mod helpers;
 pub mod import;
 pub mod types;
